+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - [#355] Excessive invalidation when scene view visibility states change
+- [#363] Reduce GC pressure caused by `ComputeContext.GetComponent`
 
 ### Changed